Add initial public code for phone box
Code prior to this was littered with secrets. It was kept mostly
in one file, with the majority of the code, so it made sense
to just blow away history and start from this point.

Secrets are now kept in the `.env` file in the root or stored as
environment variables on the production server.